requirements.txt
py-solc-x
web3
dotenv
eth-brownie
